+++ content/posts/2025-09-20-torneio-judo.mdx
@@ -2,7 +2,7 @@
 title: Alunos do IFE Conquistam Medalhas em Torneio de Judô
 date: 2025/09/20
 description: Nossa equipe de judô trouxe para casa 8 medalhas do Campeonato Regional de Artes Marciais, demonstrando técnica, disciplina e dedicação.
-tag: Esportes
+tag: [Esportes,Educação]
 author: Carlos Silva
 ---
