+++ content/posts/2025-10-15-campanha-doacao-livros.mdx
@@ -2,7 +2,7 @@
 title: Campanha de Doação de Livros Arrecada Mais de 500 Exemplares
 date: 2025/10/15
 description: Nossa biblioteca comunitária está sendo renovada graças à generosidade dos doadores. Agora temos mais opções para incentivar a leitura.
-tag: Educação
+tag: [Educação]
 author: Equipe IFE
 ---
