+++ content/posts/2025-11-10-inauguracao-nova-quadra.mdx
@@ -2,7 +2,7 @@
 title: IFE Inaugura Nova Quadra Poliesportiva
 date: 2025/11/10
 description: Com a nova quadra, ampliamos as oportunidades para mais de 200 crianças praticarem esportes em um espaço adequado e seguro.
-tag: Infraestrutura
+tag: [Infraestrutura]
 author: Equipe IFE
 ---
